fix: Move argparse functionality to cli.py
